+++ pages/client/client_list.vue
@@ -2,18 +2,7 @@
 	<block v-if="clientList.length > 0">
 		<view class="wrap">
 			<block v-for="(client, index) in clientList" :key="client._id">
-				<view class="client-item">
-					<view class="top-wrap">
-						<text class="title">{{ client.clientName }}</text>
-						<view class="tag-wrap">
-							<u-tag text="删除" type="error" @click="onClickDelete(index)"></u-tag>
-							<u-tag text="编辑" @click="onClickEdit(index)"></u-tag>
-						</view>
-					</view>
-					<text class="sub-title">报价点位: {{ client.quotationPoint }}</text>
-					<text class="sub-title">所属公司: {{ client.clientCompany }}</text>
-					<text class="sub-title">备注: {{ client.mark.length > 0 ? client.mark : '无' }}</text>
-				</view>
+				<client-item :client="client" :index="index" @delete="onClickDelete"></client-item>
 			</block>
 			<view class="add-wrap" @click="onClickAddButton">
 				<u-icon name="plus" color="#ffffff" size="60"></u-icon>
@@ -32,6 +21,7 @@
 </template>
 
 <script>
+	import ClientItem from './client_item.vue'
 	import {
 		mapGetters,
 		mapMutations
@@ -42,6 +32,9 @@
 				showDeleteModal: false,
 				deleteIndex: null
 			};
+		},
+		components: {
+			ClientItem
 		},
 		onLoad() {},
 		computed: {
@@ -65,24 +58,32 @@
 			},
 			onClickConfirmDelete() {
 				let self = this
-				uniCloud.callFunction({
-					name: 'client',
-					data: {
-						type: 'delete',
-						client: self.clientList[self.deleteIndex]
-					}
-				}).then(res => {
-					self.showDeleteModal = false
-					if (res.result) {
-						self.DELETECLIENT({
-							index: self.deleteIndex
-						})
-						self.$refs.uToast.show({
-							title: '删除成功',
-							type: 'success'
-						})
-					}
-				})
+				const client = self.clientList[self.deleteIndex]
+				if (client.openid === self.currentUser.openid) {
+					uniCloud.callFunction({
+						name: 'client',
+						data: {
+							type: 'delete',
+							client: self.clientList[self.deleteIndex]
+						}
+					}).then(res => {
+						self.showDeleteModal = false
+						if (res.result) {
+							self.DELETECLIENT({
+								index: self.deleteIndex
+							})
+							self.$refs.uToast.show({
+								title: '删除成功',
+								type: 'success'
+							})
+						}
+					})
+				} else {
+					self.$refs.uToast.show({
+						title: '无权删除',
+						type: 'error'
+					})
+				}
 			}
 		},
 	}
@@ -115,39 +116,5 @@
 		justify-content: center;
 	}
 
-	.client-item {
-		margin: 15px;
-		padding: 15px;
-		border-radius: 5px;
-		box-shadow: 1px 1px 5px 2px $u-border-color;
-		display: flex;
-		flex-direction: column;
-	}
-
-	.client-item .top-wrap {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-	}
-
-	.client-item .top-wrap .tag-wrap {
-		display: flex;
-		align-items: center;
-	}
-
-	.tag-wrap u-tag {
-		margin-left: 5px;
-	}
-
-	.client-item .title {
-		font-size: 14px;
-		font-weight: bold;
-		color: $u-main-color;
-	}
-
-	.client-item .sub-title {
-		margin-top: 5px;
-		font-size: 14px;
-		color: $u-content-color;
-	}
+	
 </style>
